requirements.txt
Flask==2.0.1
Werkzeug==2.0.2  
python-dotenv